rename usmqe/base to usmqe/web
