+++ entry_points.txt
@@ -1,3 +1,3 @@
 [usmqe.application_collections]
-users = usmqe.base.application.entities.users:UsersCollection
-clusters = usmqe.base.application.entities.clusters:ClustersCollection
+users = usmqe.web.application.entities.users:UsersCollection
+clusters = usmqe.web.application.entities.clusters:ClustersCollection
