+++ usmqe/web/application/views/user.py
@@ -2,8 +2,8 @@ from widgetastic.widget import Text, TextInput, Table
 # from widgetastic.exceptions import NoSuchElementException
 # from widgetastic.widget import GenericLocatorWidget, ParametrizedLocator, ParametrizedView
 
-from usmqe.base.application.views.common import BaseLoggedInView
-from usmqe.base.application.widgets import Kebab
+from usmqe.web.application.views.common import BaseLoggedInView
+from usmqe.web.application.widgets import Kebab
 from taretto.ui.patternfly import Button, Dropdown
