+++ usmqe/web/application/views/volume.py
@@ -1,6 +1,6 @@
 from widgetastic.widget import Text
 
-from usmqe.base.application.views.common import BaseClusterSpecifiedView
+from usmqe.web.application.views.common import BaseClusterSpecifiedView
 from taretto.ui.patternfly import Button
 from widgetastic.widget import ParametrizedLocator, ParametrizedView
